Fix warnings about WPILib version
Opening this project in VSCode with WPILib v2024.2.1 was warning "This project is not compatible with this version of the extension. Would you like to import this project into 2024?"

Updated the prefs file to remove "beta" from the WPILib version.

Also picked up an update to the license file.

And, just for grins, commit making gradlew executable, so us MacOS users don't keep having it show up on our changed lists.